moved maxCustomers to abstract class
