+++ app/src/main/java/it/gym/model/CourseTrainingBundleSpecification.java
@@ -1,55 +1,22 @@
 package it.gym.model;
 
 import com.fasterxml.jackson.annotation.JsonTypeName;
-import java.util.Objects;
-import javax.persistence.Column;
 import javax.persistence.DiscriminatorValue;
 import javax.persistence.Entity;
-import lombok.Data;
-import lombok.EqualsAndHashCode;
 import lombok.Generated;
 
 @Entity
 @DiscriminatorValue(value = "C")
 @JsonTypeName("C")
-@Data
-@EqualsAndHashCode(callSuper = true)
 @Generated // exclude coverage analysis on generated methods
 public class CourseTrainingBundleSpecification
     extends ATrainingBundleSpecification {
 
   public static final String TYPE = "C";
 
-  @Column(name = "max_customers")
-  private Integer maxCustomers;
-
-  @Override
-  public Integer getMaxCustomers() {
-    return maxCustomers;
-  }
-
-  public void setMaxCustomers(Integer maxCustomers) {
-    this.maxCustomers = maxCustomers;
-  }
-
   @Override
   public String getType() {
     return TYPE;
-  }
-
-  @Override
-  public boolean equals(Object o) {
-    if (this == o) return true;
-    if (o == null || getClass() != o.getClass()) return false;
-    if (!super.equals(o)) return false;
-    CourseTrainingBundleSpecification that =
-        (CourseTrainingBundleSpecification) o;
-    return Objects.equals(maxCustomers, that.maxCustomers);
-  }
-
-  @Override
-  public int hashCode() {
-    return Objects.hash(super.hashCode(), maxCustomers);
   }
 
   @Override
@@ -64,16 +31,10 @@ public class CourseTrainingBundleSpecification
   }
 
   @Override
-  public CourseTrainingBundleSpecification eager() {
-    super.eager();
-    return this;
-  }
-
-  @Override
   public String toString() {
     return "CourseTrainingBundleSpecification{"
         + "maxCustomers="
-        + maxCustomers
+        + getMaxCustomers()
         + '}';
   }
 }
