+++ web/src/app/shared/bundle-specs/bundle-spec-modal.component.html
@@ -53,6 +53,7 @@
             <label for="unlimitedDeletions" class="sr-only">Cancellazioni illimitate</label>
             <mat-checkbox id="unlimitedDeletions"
                           name="unlimitedDeletions"
+                          [checked]="unlimitedDeletions.value"
                           formControlName="unlimitedDeletions">Cancellazioni illimitate</mat-checkbox>
         </div>
 
@@ -84,7 +85,7 @@
                            [ngClass]="{ 'is-invalid': maxCustomers.touched && maxCustomers.invalid }" />
                     <mat-error *ngIf="maxCustomers.touched && maxCustomers.invalid">
                         <div *ngIf="!!maxCustomers.errors?.required">Numero massimo di clienti obbligatori</div>
-                        <div *ngIf="!!maxCustomers.errors?.min">Numero massimo di clienti deve essere superiore a 2</div>
+                        <div *ngIf="!!maxCustomers.errors?.min">Numero massimo di clienti deve essere superiore a 1</div>
                     </mat-error>
                 </mat-form-field>
             </div>
